requirements.txt
appdirs==1.4.3
csvImporter==0.1.3.2
Django==1.11.1
django-cors-headers==2.1.0
djangorestframework==3.6.3
numpy==1.12.1
packaging==16.8
pandas==0.20.1
pkg-resources==0.0.0
pyparsing==2.2.0
python-dateutil==2.6.0
pytz==2017.2
six==1.10.0
